% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogue_urls.R
\name{get_catalogue_urls}
\alias{get_catalogue_urls}
\title{Get OpenActive data catalogue URLs}
\usage{
get_catalogue_urls()
}
\value{
A list of OpenActive catalogue URLs
}
\description{
Get OpenActive data catalogue URLs
}
\examples{
catalogue_urls <- get_catalogue_urls()
}
